requirements.txt
tensorflow
psycopg2-binary
matplotlib
numpy
pandas
seaborn